Refactoring to make sure, the api package can be used in base package
