+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/api/ColorGradingStyle.java
@@ -1,0 +1,16 @@
+/**
+ * 
+ */
+package de.lexasoft.mandelbrot.api;
+
+/**
+ * The style of color grading in Mandelbrot colorization.
+ * 
+ * @author nierax
+ *
+ */
+public enum ColorGradingStyle {
+
+	LINE, CIRCLE
+
+}
