+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/api/MandelbrotCalculationProperties.java
@@ -10,7 +10,6 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-import de.lexasoft.mandelbrot.PaletteVariant;
 
 /**
  * @author nierax
