+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/api/PaletteVariant.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package de.lexasoft.mandelbrot;
+package de.lexasoft.mandelbrot.api;
 
 /**
  * Which color palette to use?
