+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/ctrl/CalculationPropertiesDTO.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 
-import de.lexasoft.mandelbrot.PaletteVariant;
+import de.lexasoft.mandelbrot.api.PaletteVariant;
 
 /**
  * @author admin
