+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/MandelbrotColorizeFactoryTest.java
@@ -14,6 +14,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import de.lexasoft.mandelbrot.api.PaletteVariant;
+
 /**
  * 
  * @author nierax
