+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/MandelbrotSingleRunnerTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package de.lexasoft.mandelbrot.api;
+package de.lexasoft.mandelbrot;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -18,10 +18,10 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import de.lexasoft.mandelbrot.MandelbrotBlackWhite;
-import de.lexasoft.mandelbrot.MandelbrotColorPalette;
-import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-import de.lexasoft.mandelbrot.PaletteVariant;
+import de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties;
+import de.lexasoft.mandelbrot.api.MandelbrotRunner;
+import de.lexasoft.mandelbrot.api.MandelbrotRunnerException;
+import de.lexasoft.mandelbrot.api.PaletteVariant;
 
 /**
  * @author nierax
@@ -54,7 +54,7 @@ class MandelbrotSingleRunnerTest {
 
 	/**
 	 * Test method for
-	 * {@link de.lexasoft.mandelbrot.api.MandelbrotSingleRunner#of(de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties)}.
+	 * {@link de.lexasoft.mandelbrot.MandelbrotSingleRunner#of(de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties)}.
 	 */
 	@Test
 	void testOfOk() {
