+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/api/MandelbrotCalculationPropertiesTest.java
@@ -25,7 +25,6 @@ import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 
 import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-import de.lexasoft.mandelbrot.PaletteVariant;
 
 /**
  * @author nierax
