+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/api/TransitionFactoryTest.java
@@ -20,7 +20,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import de.lexasoft.mandelbrot.MandelbrotPointPosition;
-import de.lexasoft.mandelbrot.PaletteVariant;
 
 /**
  * @author nierax
