+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/ctrl/CalculationPropertiesDTOTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 
-import de.lexasoft.mandelbrot.PaletteVariant;
+import de.lexasoft.mandelbrot.api.PaletteVariant;
 
 /**
  * @author nierax
