+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/ctrl/SingleDTO2PropertiesMapperTest.java
@@ -13,8 +13,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import de.lexasoft.mandelbrot.PaletteVariant;
 import de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties;
+import de.lexasoft.mandelbrot.api.PaletteVariant;
 
 /**
  * @author nierax
